Renamer Fullmektige -> fullmektigeGammel og flytter inn i egen mappe
