+++ src/felleskomponenter/menypanel/menypunkter/fullmektig/gammel/fullmektigRedigerer.less
