+++ src/felleskomponenter/menypanel/menypunkter/fullmektig/gammel/fullmektige.less
@@ -1,4 +1,4 @@
-@import "../../../../constants";
+@import "../../../../../constants";
 
 .fullmektige {
   .enkelt__fullmektig {
